+++ packages/atelier/button/LoadingIcon.tsx
@@ -6,36 +6,32 @@ import { buttonPrefixCls } from './constants';
 import type { LoadingIconProps } from './types';
 
 import { EXITED, useTransitionStatus } from '../transition';
-import type { Variants } from '../transition';
 
-const variants: Variants<{ iconOnly: boolean }> = {
-  enter: custom => ({
-    opacity: 1,
-    scale: 1,
-    width: '1em',
-    marginRight: custom?.iconOnly ? 0 : 8,
-  }),
-  exit: {
-    opacity: 0,
-    scale: 0,
-    width: 0,
-    marginRight: 0,
-  },
-};
-
-/** 渲染按钮内部的加载图标。 */
+/**
+ * 渲染按钮内部的加载图标。
+ */
 const LoadingIcon: FC<LoadingIconProps> = props => {
   const { className, iconOnly = false, loading = false, style } = props;
 
   const { handleRef, isMounted, status } = useTransitionStatus({
-    custom: { iconOnly },
-    enter: variants.enter,
-    exit: variants.exit,
+    appear: true,
+    enter: {
+      opacity: 1,
+      scale: 1,
+      width: 'auto',
+      marginRight: iconOnly ? 0 : 8,
+    },
+    exit: {
+      opacity: 0,
+      scale: 0,
+      width: 0,
+      marginRight: 0,
+    },
     in: loading,
     unmountOnExit: true,
   });
 
-  if (!isMounted) return;
+  if (!isMounted) return null;
 
   return (
     <span
@@ -43,8 +39,8 @@ const LoadingIcon: FC<LoadingIconProps> = props => {
       className={clsx(`${buttonPrefixCls}-loading-icon`, className)}
       ref={handleRef}
       style={{
+        visibility: status === EXITED && !loading ? 'hidden' : undefined,
         opacity: 0,
-        visibility: status === EXITED ? 'hidden' : undefined,
         width: 0,
         ...style,
       }}