+++ packages/atelier/button/constants.ts
@@ -1,4 +1,6 @@
 import { getPrefixCls } from '../utils/getPrefixCls';
 
-/** Button 组件的 CSS 类名前缀。 */
+/**
+ * Button 组件的 CSS 类名前缀。
+ */
 export const buttonPrefixCls = getPrefixCls('button');