+++ packages/atelier/button/index.ts
@@ -1,2 +1,2 @@
 export { Button } from './Button';
-export type { ButtonHTMLType, ButtonProps, ButtonShape, ButtonStatus, ButtonType } from './types';
+export type { ButtonProps, ButtonShape, ButtonVariant } from './types';